Created base factory class to handle list deserialisation
The deserialisation of an entity from the JSON returned by the API is
unique to the entity, but the process of deserialising a collection is
not - it's the same for each. I have therefore moved the
from_json_list() method into a base class from which the Factory classes
inherit.