Updated setup.py
As part of the break away from the original project I have updated the
setup.py file. I don't want to pretend that I did this all myself, so
need to acknowledge the original project somehow, but it's now
sufficiently diverged to make it reasonable to start to update the
package settings.